requirements.txt
Flask
langchain
langchain-core
langchain-openai
requests
gunicorn
pytest
python-dotenv
psycopg2-binary